requirements.txt
feedparser==6.0.11
pandas==2.2.2
pyyaml==6.0.2
folium==0.17.0
python-dateutil==2.9.0.post0
requests==2.32.3
